requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
scikit-learn>=1.4.0
plotly>=5.20.0
pandas>=2.2.0
numpy>=1.26.0
spacy>=3.7.0
requests>=2.31.0
datasets>=2.18.0
python-dotenv>=1.0.0
torch>=2.0.0
pytest>=8.0.0
